"add Shape Area and tests + README entry"
