+++ src/main/java/Intro/Add.java
@@ -1,4 +1,4 @@
-package TheJourneyBegins;
+package Intro;
 
 public class Add {
     int solution(int param1, int param2) {
