+++ src/main/java/Intro/CenturyFromYear.java
@@ -1,0 +1,7 @@
+package Intro;
+
+public class CenturyFromYear {
+    int solution(int year) {
+        return (int) Math.ceil((double)year/100);
+    }
+}
