--- src/main/java/TheJourneyBegins/CenturyFromYear.java
@@ -1,7 +1,0 @@
-package TheJourneyBegins;
-
-public class CenturyFromYear {
-    int solution(int year) {
-        return (int) Math.ceil((double)year/100);
-    }
-}
