+++ src/test/java/Intro/TheJourneyBeginsTest.java
@@ -1,6 +1,5 @@
-package TheJourneyBegins;
+package Intro;
 
-import org.checkerframework.checker.units.qual.C;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
